na9es modif ta3 fournisseur

+++ Gestion_Stock/src/riadh/gestion/dao/FournisseurDAO.java
@@ -8,8 +8,8 @@ public interface FournisseurDAO {
 	
 	public void add(Fournisseur f);
 	public Fournisseur edite(Fournisseur f);
-	public void delete(long id);
+	public void delete(int id);
 	public List<Fournisseur> finAll();
-	public Fournisseur finById(long id);
+	public Fournisseur finById(int id);
 
 }
