+++ Gestion_Stock/src/riadh/gestion/service/FournisseurService.java
@@ -8,8 +8,8 @@ public interface FournisseurService {
 	
 	public void add(Fournisseur f);
 	public Fournisseur edite(Fournisseur f);
-	public void delete(long id);
+	public void delete(int id);
 	public List<Fournisseur> finAll();
-	public Fournisseur finById(long id);
+	public Fournisseur finById(int id);
 
 }
